Rename Match to LegacyMatch as the former is php 8 reserved keyword
