+++ database/factories/Chat/ChannelFactory.php
@@ -3,7 +3,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 use App\Models\Chat\Channel;
-use App\Models\Match\Match;
+use App\Models\LegacyMatch\LegacyMatch;
 
 $factory->define(Channel::class, function (Faker\Generator $faker) {
     return [
@@ -25,7 +25,7 @@ $factory->state(Channel::class, 'pm', function (Faker\Generator $faker) {
 });
 
 $factory->state(Channel::class, 'tourney', function (Faker\Generator $faker) {
-    $match = factory(Match::class)->states('tourney')->create();
+    $match = factory(LegacyMatch::class)->states('tourney')->create();
 
     return [
         'name' => "#mp_{$match->match_id}",
