+++ database/factories/LegacyMatch/GameFactory.php
@@ -3,7 +3,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-use App\Models\Match\Game;
+use App\Models\LegacyMatch\Game;
 use Carbon\Carbon;
 
 $factory->define(Game::class, function (Faker\Generator $faker) {
