+++ database/factories/LegacyMatch/LegacyMatchFactory.php
@@ -3,9 +3,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-use App\Models\Match\Match;
+use App\Models\LegacyMatch\LegacyMatch;
 
-$factory->define(Match::class, function (Faker\Generator $faker) {
+$factory->define(LegacyMatch::class, function (Faker\Generator $faker) {
     return [
         'name' => function () use ($faker) {
             return $faker->sentence();
@@ -15,13 +15,13 @@ $factory->define(Match::class, function (Faker\Generator $faker) {
     ];
 });
 
-$factory->state(Match::class, 'private', function (Faker\Generator $faker) {
+$factory->state(LegacyMatch::class, 'private', function (Faker\Generator $faker) {
     return [
         'private' => 1,
     ];
 });
 
-$factory->state(Match::class, 'tourney', function (Faker\Generator $faker) {
+$factory->state(LegacyMatch::class, 'tourney', function (Faker\Generator $faker) {
     return [
         'keep_forever' => 1,
     ];
